Connect New Tab Page Duck.ai shortcut setting to NTP (#1469)
Task/Issue URL: https://app.asana.com/1/137249556945/project/1201048563534612/task/1210867989812277?focus=true
CC: @tomasstrba

Description:
This change connects native Duck.ai settings to New Tab Page Omnibar configuration to hide the Duck.ai
input prompt if the user wishes so.
NewTabPageOmnibarModeProvider was renamed to NewTabPageOmnibarConfigProvider (because config
is now more than just the mode). The provider was updated with AI chat settings to be able to pass
AI settings to NTP.
AIChatMenuConfiguration was moved to App Delegate and it’s injected to all classess using it.