Fix: Infinite Spinner Glitch (#2578)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/72649045549333/task/1211992423969266?focus=true
Tech Design URL:
CC:

### Description
In this PR we're fixing a glitch that caused the new Tab Bar Spinner to
go on indefinitely.

### Details
The legacy Address Bar Activity Indicator relied on
`TabViewModel.isLoading`, whereas the new Tab Spinner relies on
`Tab.isLoading`.

I've managed to verify that when this glitch occurs, `webView.isLoading`
was still `true`.
Reason why this didn't happen with the old Activity Indicator is that we
were marking `TabViewModel.isLoading = false` when the `Page DOM Loaded`
Event was posted.

This changeset brings the new implementation on par with the old
behavior.