Fix: Tab Item Title Overlap (#2951)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1211150618152277/task/1212489595848548?focus=true
Tech Design URL:
CC:

### Description
In this PR we're fixing a layout issue affecting the Tab's Title, when
in compact mode.

#### Details
Prior to the changes introduced via the `tabProgressIndicator` flag, the
FaviconView wouldn't be rendered whenever a given website lacked a
Favicon Image.

When the `tabProgressIndicator` flag is enabled, the `TabFaviconView`
component will always be visible, since we'll display a placeholder view
(where the favicon normally goes).

We've adjusted the logic that conceals the Tab Item's Title, so that it
accounts for this difference, and thus, prevents overlapping labels due
to space limitations.